feat(user-CRUD): add route to find user by email using query
This commit adds a new route to the application that allows users to be found by email using a query. The route accepts an email address as a parameter and returns the user entity if a match is found. If no match is found, the route returns a 404 status code.