feat(serialize-interceptor): Add serialize interceptor to remove password from response
We added this new feature to remove the password from the response that is sent back to the user. The serialize interceptor allows us to modify the response object before it is sent back to the client, which we used to remove the password property from the response data.

This change helps to protect the security of our users by ensuring that sensitive information is not accidentally exposed.